+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2019-01-30
+### Removed
+- Removed default `by_query` scope from `Folio::Filterable` - use custom `PgSearch` instead!
+
 ## 2019-01-29
 ### Added
 - Added autocomplete input `f.input :field, as: :autocomplete, collection: ['foo', 'bar']`